praca_inz-main/requirements.txt
azure-core==1.32.0
azure-storage-blob==12.24.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
eralchemy==1.5.0
Flask==3.1.0
fonttools==4.55.3
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
numpy==2.2.1
packaging==24.2
pillow==11.0.0
psycopg2==2.9.10
pycparser==2.22
pyparsing==3.2.1
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
